Change how secrets are generated in dev and prod
In dev, when the server was restarted (e.g. because of a file change), new jwt secrets were generated, requiring everyone to login again.

That isn't really a problem in production, but in dev it was annoying.

So, now it only regenerates secrets in production. In development, it uses predefined, insecure, secrets.

This will not affect anything else.